+++ src/imp/libc/syscalls.rs
@@ -10,6 +10,8 @@ pub(crate) use super::net::syscalls::*;
 pub(crate) use super::process::syscalls::*;
 #[cfg(not(windows))]
 pub(crate) use super::time::syscalls::*;
+#[cfg(not(windows))]
+pub(crate) use super::thread::syscalls::*;
 
 #[cfg(any(windows, target_os = "android", target_os = "linux"))]
 use super::c;
