+++ src/imp/libc/thread/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(not(windows))]
+pub(crate) mod syscalls;
