+++ src/imp/libc/thread/syscalls.rs
@@ -1,0 +1,78 @@
+use super::super::c;
+use super::super::conv::ret;
+use super::super::time::Timespec;
+use core::mem::MaybeUninit;
+use crate::io;
+#[cfg(not(any(
+    target_os = "dragonfly",
+    target_os = "freebsd",
+    target_os = "emscripten",
+    target_os = "ios",
+    target_os = "macos",
+    target_os = "openbsd",
+    target_os = "redox",
+    target_os = "wasi",
+)))]
+use {
+    core::ptr::null_mut,
+    super::super::time::ClockId,
+};
+#[cfg(not(target_os = "redox"))]
+use crate::thread::NanosleepRelativeResult;
+
+#[cfg(not(any(
+    target_os = "dragonfly",
+    target_os = "emscripten",
+    target_os = "freebsd", // FreeBSD 12 has clock_nanosleep, but libc targets FreeBSD 11.
+    target_os = "ios",
+    target_os = "macos",
+    target_os = "openbsd",
+    target_os = "redox",
+    target_os = "wasi",
+)))]
+#[inline]
+pub(crate) fn clock_nanosleep_relative(id: ClockId, request: &Timespec) -> NanosleepRelativeResult {
+    let mut remain = MaybeUninit::<Timespec>::uninit();
+    let flags = 0;
+    unsafe {
+        match c::clock_nanosleep(id as c::clockid_t, flags, request, remain.as_mut_ptr()) {
+            0 => NanosleepRelativeResult::Ok,
+            err if err == io::Error::INTR.0 => {
+                NanosleepRelativeResult::Interrupted(remain.assume_init())
+            }
+            err => NanosleepRelativeResult::Err(io::Error(err)),
+        }
+    }
+}
+
+#[cfg(not(any(
+    target_os = "dragonfly",
+    target_os = "freebsd", // FreeBSD 12 has clock_nanosleep, but libc targets FreeBSD 11.
+    target_os = "emscripten",
+    target_os = "ios",
+    target_os = "macos",
+    target_os = "openbsd",
+    target_os = "redox",
+    target_os = "wasi",
+)))]
+#[inline]
+pub(crate) fn clock_nanosleep_absolute(id: ClockId, request: &Timespec) -> io::Result<()> {
+    let flags = c::TIMER_ABSTIME;
+    match unsafe { c::clock_nanosleep(id as c::clockid_t, flags, request, null_mut()) } {
+        0 => Ok(()),
+        err => Err(io::Error(err)),
+    }
+}
+
+#[cfg(not(target_os = "redox"))]
+#[inline]
+pub(crate) fn nanosleep(request: &Timespec) -> NanosleepRelativeResult {
+    let mut remain = MaybeUninit::<Timespec>::uninit();
+    unsafe {
+        match ret(c::nanosleep(request, remain.as_mut_ptr())) {
+            Ok(()) => NanosleepRelativeResult::Ok,
+            Err(io::Error::INTR) => NanosleepRelativeResult::Interrupted(remain.assume_init()),
+            Err(err) => NanosleepRelativeResult::Err(err),
+        }
+    }
+}
