+++ tests/thread/main.rs
@@ -1,4 +1,6 @@
 #![cfg(not(windows))]
 
+#[cfg(not(any(target_os = "redox")))]
+mod clocks;
 #[cfg(any(target_os = "android", target_os = "linux"))]
 mod id;
